+++ private/bufpkg/bufmodule/bufmodulebuild/module_bucket_builder_test.go
@@ -355,7 +355,7 @@ lint:
 		bufmoduleconfig.ExternalConfigV1{},
 	)
 	require.NoError(t, err)
-	module, err := BuildForBucket(
+	module, err := NewModuleBucketBuilder().BuildForBucket(
 		ctx,
 		bucket,
 		config,
@@ -408,7 +408,7 @@ func testBucketGetFileInfos(
 		storageos.ReadWriteBucketWithSymlinksIfSupported(),
 	)
 	require.NoError(t, err)
-	module, err := BuildForBucket(
+	module, err := NewModuleBucketBuilder().BuildForBucket(
 		context.Background(),
 		readWriteBucket,
 		config,
@@ -431,7 +431,7 @@ func testBucketGetFileInfos(
 			require.NoError(t, err)
 			bucketRelPaths[i] = bucketRelPath
 		}
-		module, err := BuildForBucket(
+		module, err := NewModuleBucketBuilder().BuildForBucket(
 			context.Background(),
 			readWriteBucket,
 			config,
@@ -461,7 +461,7 @@ func testBucketGetAllFileInfosError(
 		storageos.ReadWriteBucketWithSymlinksIfSupported(),
 	)
 	require.NoError(t, err)
-	module, err := BuildForBucket(
+	module, err := NewModuleBucketBuilder().BuildForBucket(
 		context.Background(),
 		readWriteBucket,
 		config,
@@ -496,7 +496,7 @@ func testBucketGetFileInfosForExternalPathsError(
 		require.NoError(t, err)
 		bucketRelPaths[i] = bucketRelPath
 	}
-	_, err = BuildForBucket(
+	_, err = NewModuleBucketBuilder().BuildForBucket(
 		context.Background(),
 		readWriteBucket,
 		config,
@@ -522,7 +522,7 @@ func testDocumentationBucket(
 		bufmoduleconfig.ExternalConfigV1{},
 	)
 	require.NoError(t, err)
-	module, err := BuildForBucket(
+	module, err := NewModuleBucketBuilder().BuildForBucket(
 		context.Background(),
 		readWriteBucket,
 		config,
@@ -556,7 +556,7 @@ func testLicenseBucket(
 		bufmoduleconfig.ExternalConfigV1{},
 	)
 	require.NoError(t, err)
-	module, err := BuildForBucket(
+	module, err := NewModuleBucketBuilder().BuildForBucket(
 		context.Background(),
 		readWriteBucket,
 		config,