+++ private/bufpkg/bufmodule/bufmoduleref/bufmoduleref_test.go
@@ -158,9 +158,8 @@ func bucketWithBufLock(t *testing.T, pin ModulePin) storage.ReadWriteBucket {
 			},
 		},
 	}
-	bucket, err := storagemem.NewReadWriteBucketWithOptions()
-	require.NoError(t, err)
-	err = buflock.WriteConfig(context.Background(), bucket, bufLock)
+	bucket := storagemem.NewReadWriteBucket()
+	err := buflock.WriteConfig(context.Background(), bucket, bufLock)
 	require.NoError(t, err)
 	return bucket
 }